Move S044 analysis into facts
